backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.0,<3.0
xgboost
scikit-learn
supabase
apscheduler
httpx
python-dotenv
